Add service redeploy
